Compile cross VM at high debug for source info
This probably slows things down a little bit, but it has an
important advantage: assuming the host Lisp is decent, it can keep
enough information that the bytecode module and current IP can be
extracted from a host frame. And with that you can get source info
and whatever else. I wrote a proof of concept of this in SBCL, but
you need to use undocumented/fragile internals, so I won't be
including it in Maclina at least for the moment.