Don't use host literals for externalization tests
This has been masking a bug. The test supposedly tests vectors with
fill pointers. However, because the vector with fill pointer was a
literal, it was being dumped by the host. Some hosts, e.g. SBCL,
will simplify complex arrays. So Maclina never actually got an
array with fill pointer to test, and so I didn't notice that our
externalization of arrays with fill pointers is broken.