Track source positions and use them in errors
This is mostly based on the approach that works for Clasp (which
is itself based on SBCL's) with a few small changes.
In Clasp there is a special print-compiler-condition function used
by the compiler, whereas here compiler conditions print their
source location as part of their normal print-object behavior. This
allows source information to be obvious even outside of the compiler.
Secondly the source location is just printed through the normal
printer; a client could define their own method on
make-parse-result, hypothetically.