Support some "nesting" in global environments
This lets you make a sort of proxy environment over the real
global environment if you're so inclined. I've found this to be
useful for the maybe wrongheaded approach of using the compiler as
a code walker: If I want to see if some code uses a function, I
can compile it in a proxy environment that shadows that name, and
look for an unknown operator warning.