Generate map info in correct order
Nested functions broke the invariant that successive infos would
have nondecreasing start points.

An alternative way to do this would be to continue to push infos
to a module vector instead of the new function vectors, and then
sort at the end. But I'd like to avoid sneaking in supralinear
behavior as much as I can manage.

Changes the interface slightly in that cmodule-pc-map disappears,
but I think the new interface is better. The less outsiders can
access of compiler modules the better.