Remove extraneous null bytes from linked FASLs
Since instruction bytes are just read into a simple vector that's
just resized to be big, extra space on the end could be copied
into linked FASLs. Technically they are uninitialized and not
necessarily zero, but that just makes it worse.

Maclina's FASL loader would complain about this but Clasp's does
not, so I didn't notice. It would just ignore extra bytes on the
end. However, if we link linked FASLs, the null bytes would
dutifully be copied into the new FASL and there cause a problem as
more instructions may actually be expected.