native file compilation: tolerate a function not appearing
Originally I was going for an ideal of every bytecode function
being compiled even if it's optimized away by the native code,
since it could still be called from bytecode, but maybe that's
silly... a function can be truly useless if it's inlinable: there's
no way for it to be called except from the also-compiled function
it was inlined into. So we don't need to enforce that ideal.