CFASLs
These are FASLs that have the _compile-time_ side effects of the
file, rather than load-time as with normal FASLs. These are useful
to have for some build purposes, e.g. to build part of a system in
one image and then continue the build in another image.

Experimental. Pass :OUTPUT-CFASL to compile-file, etc. It should
be a pathname.